chore: reorganize util
