+++ internal/validationutil/result_db.go
@@ -3,30 +3,10 @@ package validationutil
 import (
 	"errors"
 	"fmt"
-	"net/mail"
 	"regexp"
-	"strconv"
 
 	"github.com/Project-Quantum-Workspace/QuantumLab/model"
 )
-
-// validate the auto increment ID
-func ValidateID(str string) (uint, error) {
-	id, err := strconv.Atoi(str)
-	if err != nil {
-		return 0, err
-	}
-	if id <= 0 {
-		return 0, errors.New("invalid id")
-	}
-	return uint(id), nil
-}
-
-// validate email address
-func ValidateEmail(email string) bool {
-	_, err := mail.ParseAddress(email)
-	return err == nil
-}
 
 func ValidateTableCreationRequest(table *model.CreateTableRequest) error {
 	columnCheck := validateColumnCount(table.ColumnCount, table.ColumnName,
